refactor: 컨트롤러에 ProductService 추가
